Remove legacy web component paths
