+++ apps/web/app/layout.tsx
@@ -1,24 +1,2 @@
-import type { Metadata } from 'next';
-import type { ReactNode } from 'react';
-import { Inter } from 'next/font/google';
-import './globals.css';
-
-const inter = Inter({ subsets: ['latin'] });
-
-export const metadata: Metadata = {
-  title: 'E-CO Experience',
-  description:
-    'Immersive Ether Computing interface for SymbolCast and Q++ workflows.',
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: ReactNode;
-}) {
-  return (
-    <html lang="en">
-      <body className={`${inter.className} bg-black text-white`}>{children}</body>
-    </html>
-  );
-}
+export { metadata } from '../frontend/app/layout';
+export { default } from '../frontend/app/layout';
