+++ apps/web/frontend/README.md
@@ -1,0 +1,5 @@
+# Frontend
+
+This directory houses the client-facing pieces of the E-CO web shell. The Next.js `app` router entrypoints live in `app/` and re-exported wrappers in the workspace root keep the framework happy. Reusable UI building blocks – such as the desktop shell, command palette, and dashboards – sit in `components/` alongside their co-located tests.
+
+Expect to find presentation code, hooks, and other browser-only logic here.
