+++ apps/web/frontend/app/globals.css
