+++ apps/web/frontend/components/CommandPalette.tsx
