+++ apps/web/frontend/components/ForegroundOverlay.tsx
