+++ apps/web/frontend/components/Taskbar.tsx
