+++ apps/web/frontend/components/__tests__/CommandPalette.test.tsx
