+++ apps/web/frontend/components/__tests__/LauncherMenu.test.tsx
