+++ apps/web/frontend/components/__tests__/Taskbar.test.tsx
