+++ apps/web/next.config.js
@@ -1,3 +1,5 @@
+const path = require('node:path');
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   experimental: {
@@ -7,6 +9,12 @@ const nextConfig = {
   webpack: (config) => {
     config.experiments = config.experiments || {};
     config.experiments.topLevelAwait = true;
+    config.resolve = config.resolve || {};
+    config.resolve.alias = {
+      ...(config.resolve.alias || {}),
+      '@frontend': path.resolve(__dirname, 'frontend'),
+      '@backend': path.resolve(__dirname, 'backend'),
+    };
     return config;
   },
 };
